Refactor utils.nzb to plugins.downloaders.nzb.utils
