+++ core/utils/__init__.py
@@ -20,7 +20,6 @@ from core.utils.links import copy_link, replace_links
 from core.utils.naming import clean_file_name, is_sample, sanitize_name
 from core.utils.network import find_download, server_responding, test_connection, wake_on_lan, wake_up
 from core.utils.notifications import plex_update
-from core.utils.nzbs import get_nzoid, report_nzb
 from core.utils.parsers import (
     parse_args,
     parse_deluge,
